+++ src/app/app-routing.ts
@@ -1,6 +1,5 @@
 import { Routes } from '@angular/router';
 import { AutoLoginPartialRoutesGuard } from 'angular-auth-oidc-client';
-import { CredentialOfferOnboardingComponent } from './features/credential-offer-onboarding/credential-offer-onboarding.component';
 import { OnboardingPolicy } from "./core/policies/onboarding-policy";
 
 export const routes: Routes = [
@@ -20,7 +19,7 @@ export const routes: Routes = [
   {
     path: 'organization/credentials/create',
     loadChildren: () =>
-      import('./features/credential-issuance/credential-issuance-routes').then(
+      import('./features/credential-issuance/credential-issuance.routes').then(
         (m) => m.default
       ),
     canActivate: [AutoLoginPartialRoutesGuard, OnboardingPolicy],
@@ -28,17 +27,13 @@ export const routes: Routes = [
   {
     path: 'organization/credentials/create2/:id',
     loadChildren: () =>
-      import('./features/credential-issuance/credential-issuance-routes').then(
+      import('./features/credential-issuance/credential-issuance.routes').then(
         (m) => m.default
       ),
     canActivate: [AutoLoginPartialRoutesGuard, OnboardingPolicy],
   },
   {
     path: 'credential-offer',
-    component:CredentialOfferOnboardingComponent,
-  },
-  {
-    path: 'credential-offer-detail',
     loadChildren: () =>
       import('./features/credential-offer/credential-offer.routes').then(
         (m) => m.default