+++ src/app/core/services/credential-procedure.service.ts
@@ -1,11 +1,12 @@
 import { inject, Injectable } from '@angular/core';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { Observable, throwError } from 'rxjs';
-import { catchError, map } from 'rxjs/operators';
+import { catchError } from 'rxjs/operators';
 import { environment } from 'src/environments/environment';
 import { ProcedureRequest } from '../models/dto/procedure-request.dto';
 import { ProcedureResponse } from "../models/dto/procedure-response.dto";
 import { LearCredentialEmployeeDataDetail } from "../models/dto/lear-credential-employee-data-detail.dto";
+import { CredentialOfferResponse } from '../models/dto/credential-offer-response';
 
 @Injectable({
   providedIn: 'root'
@@ -45,18 +46,17 @@ export class CredentialProcedureService {
     );
   }
 
-  public getCredentialOffer(transactionCode: string): Observable<string> {
-    return this.http.get(`${this.credentialOfferUrl}/transaction-code/${transactionCode}`, { responseType: 'text' }).pipe(
-      map(response => {
-        try {
-          const jsonResponse = JSON.parse(response);
-          return jsonResponse.qrCode || response;
-        } catch (e) {
-          return response;
-        }
-      }),
+  public getCredentialOfferByTransactionCode(transactionCode: string): Observable<CredentialOfferResponse> {
+    console.info('Getting credential offer by transaction code: ' + transactionCode);
+    return this.http.get<CredentialOfferResponse>(`${this.credentialOfferUrl}/transaction-code/${transactionCode}`).pipe(
       catchError(this.handleError)
     );
+  }
+
+  public getCredentialOfferByCTransactionCode(cTransactionCode: string): Observable<CredentialOfferResponse> {
+    console.info('Refreshing QR code: getting credential offer by c-transaction code: ' + cTransactionCode);
+    return this.http.get<CredentialOfferResponse>(`${this.credentialOfferUrl}/c-transaction-code/${cTransactionCode}`).pipe(
+      catchError(this.handleError));
   }
 
   private handleError(error: HttpErrorResponse) {
@@ -66,8 +66,8 @@ export class CredentialProcedureService {
     } else {
       errorMessage = `Server-side error: ${error.status} ${error.message}`;
     }
-    console.error('Error response body:', error.error);
-    return throwError(()=>errorMessage);
+    console.error('Error response body:', errorMessage);
+    return throwError(()=>error);
   }
 
   /**