+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [1.4.0](https://github.com/in2workspace/in2-issuer-ui/releases/tag/v1.4.0)
+### Added
+- Same-device flow: user can get credential with a signel device, without need to scan QR
+- Stepper to get credential offer
+
 ### [1.3.0](https://github.com/in2workspace/in2-issuer-ui/releases/tag/v1.3.0)
 ### Changed
 - Add new attribute to the credentials view